Add scroll down indicator
It was not abundantly clear on first visit that there was anything more to the site than the home page (aka it was not clear that you could scroll). A downward pointing arrow now appears after 4 seconds if the user does not scroll, and is animated. This disappears as soon as the user begins scrolling.