Implement new project display layout
Previously, the project displays on large screens (1024px and wider) consisted of a large project displays section followed by a small project displays section below it. Now the section shows both of them side by side. This helps the visual presentation of the sections when one of the small project displays are open.

One of the lines of css required to implement this did not compile even though the usage of calc was correct syntactically. The content width is currently hard-coded in until a fix is pushed out by the offending package.